projeto_gerenciador_planilhas/requirements.txt
# Google Sheets API e autenticação
gspread==5.2.0
oauth2client==4.1.3

# Manipulação de planilhas e dados
pandas==1.3.5
openpyxl==3.0.9
xlsxwriter==1.4.4

# Visualização e Dashboard
seaborn==0.11.2
matplotlib==3.7.2
PyQt5==5.15.4  # Ou PySide2 caso prefira

# Logs e utilitários de data e hora
python-dateutil==2.8.2
